+++ src/other/libregex/regex.h
@@ -56,7 +56,7 @@
 /* ugly hack to make non-MS windows compilers work */
 #if defined(_WIN32) && !defined(_MSC_VER) && !defined(_OFF_T_)
 #define _OFF_T_
-typedef long off_t;
+typedef ssize_t off_t;
 #endif
 
 typedef off_t regoff_t;
